+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/registry/PolarisRegistrationTest.java
@@ -46,7 +46,6 @@ import static com.tencent.polaris.test.common.Consts.SERVICE_PROVIDER;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 /**
  * Test for {@link PolarisRegistration}.
@@ -82,7 +81,6 @@ public class PolarisRegistrationTest {
 
 		// mock NacosContextProperties
 		nacosContextProperties = mock(NacosContextProperties.class);
-		doReturn(true).when(nacosContextProperties).isEnabled();
 		doReturn(true).when(nacosContextProperties).isRegisterEnabled();
 		doReturn("/").when(nacosContextProperties).getContextPath();
 		doReturn("cluster").when(nacosContextProperties).getClusterName();
@@ -179,7 +177,7 @@ public class PolarisRegistrationTest {
 		Map<String, String> metadata = polarisRegistration1.getMetadata();
 		assertThat(metadata).isNotNull();
 		assertThat(metadata).isNotEmpty();
-		assertThat(metadata.size()).isEqualTo(2);
+		assertThat(metadata.size()).isEqualTo(1);
 		assertThat(metadata.get("key1")).isEqualTo("value1");
 	}
 
@@ -191,25 +189,5 @@ public class PolarisRegistrationTest {
 	@Test
 	public void testToString() {
 		System.out.println(polarisRegistration1);
-	}
-
-	@Test
-	public void testGetNacosServiceId() {
-		String groupName = "group";
-		String format = "%s__%s";
-		when(nacosContextProperties.getGroup()).thenReturn(groupName);
-		String serviceId = polarisRegistration1.getServiceId();
-		assertThat(String.format(format, groupName, SERVICE_PROVIDER).equals(serviceId));
-	}
-
-	@Test
-	public void testGetNacosMetadata() {
-		String clusterName = "cluster";
-		when(nacosContextProperties.getClusterName()).thenReturn(clusterName);
-		Map<String, String> metadata = polarisRegistration1.getMetadata();
-		assertThat(metadata).isNotNull();
-		assertThat(metadata).isNotEmpty();
-		assertThat(metadata.size()).isEqualTo(2);
-		assertThat(metadata.get("nacos.cluster")).isEqualTo(clusterName);
 	}
 }